requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2024.2
Django==4.2
django-cpf==0.1.0
django-crispy-forms==2.1
django-multipleselectionfield==1.1.1
flake8==7.1.1
idna==3.7
kombu==5.4.2
mccabe==0.7.0
pillow==10.3.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.2
six==1.17.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
